+++ packages/libs/lambda-at-edge/src/default-handler.ts
@@ -158,7 +158,7 @@ const router = (
       return html.nonDynamic[normalisedUri];
     }
 
-    // Dynamic routes are matched first
+    // dynamic routes are matched first
     for (const route in allDynamicRoutes) {
       const { file, regex } = allDynamicRoutes[route];