+++ packages/compat-layers/lambda-at-edge-compat/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.1](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/next-aws-cloudfront@1.6.1-alpha.0...@sls-next/next-aws-cloudfront@1.6.1) (2021-05-03)
+
+**Note:** Version bump only for package @sls-next/next-aws-cloudfront
+
 ## [1.6.1-alpha.0](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/next-aws-cloudfront@1.6.0...@sls-next/next-aws-cloudfront@1.6.1-alpha.0) (2020-12-05)
 
 **Note:** Version bump only for package @sls-next/next-aws-cloudfront