+++ packages/deprecated/serverless-plugin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.5](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/serverless-plugin@2.6.5-alpha.4...@sls-next/serverless-plugin@2.6.5) (2021-05-03)
+
+**Note:** Version bump only for package @sls-next/serverless-plugin
+
 ## [2.6.5-alpha.4](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/serverless-plugin@2.6.5-alpha.3...@sls-next/serverless-plugin@2.6.5-alpha.4) (2021-04-21)
 
 **Note:** Version bump only for package @sls-next/serverless-plugin