+++ packages/serverless-components/domain/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/domain@1.1.1-alpha.1...@sls-next/domain@1.1.1) (2021-05-03)
+
+**Note:** Version bump only for package @sls-next/domain
+
 ## [1.1.1-alpha.1](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/domain@1.1.0...@sls-next/domain@1.1.1-alpha.1) (2020-11-13)
 
 ### Bug Fixes