+++ packages/serverless-components/serverless-cli/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.3](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/serverless-cli@1.5.3-alpha.0...@sls-next/serverless-cli@1.5.3) (2021-05-03)
+
+**Note:** Version bump only for package @sls-next/serverless-cli
+
 ## 1.5.3-alpha.0 (2021-04-15)
 
 **Note:** Version bump only for package @sls-next/serverless-cli